test: print time_t as int64

+++ test/datetime_test.c
@@ -1,11 +1,12 @@
 #include <modernlib/datetime.h>
+#include <inttypes.h>
 #include "utest.h"
 
 UTEST(DATETIME, NOW) {
     datetime now = datetime_now();
 
-    fprintf(stdout, "%ld, offset from utc %lf seconds\n", 
-            now._c11time.tv_sec,
+    fprintf(stdout, "%"PRIu64", offset from utc %lf seconds\n", 
+            (int64_t)now._c11time.tv_sec,
             timespan_as_seconds(now.offset_from_utc));
 }
